fix: 移除 Chat 模块遗留 MCP 文本层，修复 Markdown 渲染器随消息数累积变慢的问题
- chat.ts：删除旧版文本模式匹配的 MCP 调用层（executeMcpTool/handleMcpCalls/
  buildMcpSystemPrompt/buildMcpToolDefinitions）。后端原生工具调用循环上线后
  该层已匹配不到任何真实输出，但仍在每轮请求把全部工具定义以 JSON 文本塞进
  system prompt，与后端原生 tools 字段重复声明、白白多付一份 token。
- ChatMessage.vue：marked.use()/DOMPurify.addHook()/mermaid.initialize() 原先
  写在 <script setup> 顶层，每挂载一条消息都会对全局单例重复叠加扩展/钩子，
  长会话下 Markdown 解析逐条变慢、重启才恢复。整个渲染管线迁移到新模块
  utils/markdown.ts 的模块作用域，应用生命周期内只初始化一次。
- 版本号 0.2.0-beta.15 → 0.2.0-beta.16。
