+++ package.json
@@ -1,7 +1,7 @@
 {
   "name": "baiyu-aispace",
   "private": true,
-  "version": "0.2.0-beta.15",
+  "version": "0.2.0-beta.16",
   "type": "module",
   "scripts": {
     "dev": "vite",
