+++ src-tauri/tauri.conf.json
@@ -1,7 +1,7 @@
 {
   "$schema": "../node_modules/@tauri-apps/cli/config.schema.json",
   "productName": "BaiyuAISpace",
-  "version": "0.2.0-beta.15",
+  "version": "0.2.0-beta.16",
   "identifier": "com.baiyu.aispace",
   "build": {
     "frontendDist": "../dist",
