+++ src/components/ChatMessage.vue
@@ -28,23 +28,11 @@ import { computed, ref, watch, onMounted, onBeforeUnmount, nextTick } from "vue"
 // 导入 NaiveUI 组件
 import { NAvatar, NIcon, NSpin, NAlert, NTooltip } from "naive-ui";
 
-// 导入 Markdown 解析库
-import { marked } from "marked";
-
-// 导入 LaTeX 公式渲染扩展
-import markedKatex from "marked-katex-extension";
-import "katex/dist/katex.min.css";
-
-// 导入 HTML 净化库
-import DOMPurify from "dompurify";
-
-// 导入代码高亮库
-import hljs from "highlight.js";
-// 灰度高亮主题，契合黑白设计系统
-import "highlight.js/styles/grayscale.css";
-
-// 导入 Mermaid 图表库
-import mermaid from "mermaid";
+// Markdown 渲染管线。marked/KaTeX/DOMPurify/hljs/Mermaid 的全局配置都在该
+// 模块顶层完成，整个应用只执行一次——不能放回本组件的 <script setup>：
+// setup 顶层语句每挂载一条消息都会重新执行，会给全局单例反复叠加扩展/钩子，
+// 长会话下渲染越来越慢（详见 utils/markdown.ts 头部说明）。
+import { renderMarkdown, renderMermaidDiagrams } from "@/utils/markdown";
 
 // 导入消息类型
 import type { Message } from "@/stores/chat";
@@ -58,50 +46,13 @@ const props = defineProps<{
   message: Message;
 }>();
 
-// ============ Mermaid 初始化 ============
-
-mermaid.initialize({
-  startOnLoad: false,
-  // neutral 主题为灰度配色，契合黑白设计系统
-  theme: "neutral",
-  securityLevel: "loose",
-  fontFamily: '"Inter Variable", "Inter", system-ui, sans-serif',
-});
-
-// DOMPurify 默认的 URI 安全校验会把 srcdoc 这种"值不是 URL 而是一整段
-// HTML"的属性直接剥离，用 hook 强制放行 -- buildHtmlPreviewBlock 生成的
-// srcdoc 内容里的引号已经手动转义过，不依赖这里的校验来防属性逃逸。
-DOMPurify.addHook("uponSanitizeAttribute", (_node, data) => {
-  if (data.attrName === "srcdoc") {
-    data.forceKeepAttr = true;
-  }
-});
-
 // ref 指向渲染 markdown 的 DOM 节点，用于查找 Mermaid 占位元素
 const contentRef = ref<HTMLElement | null>(null);
 
-// 唯一 ID 计数器，mermaid.render() 要求每次传不同 id
-let mermaidIdCounter = 0;
-
-/** 查找并渲染当前消息内所有待渲染的 Mermaid 占位块 */
-async function renderMermaidDiagrams() {
+/** 渲染当前消息内所有待渲染的 Mermaid 占位块 */
+async function renderPendingDiagrams() {
   if (!contentRef.value) return;
-  const pending = contentRef.value.querySelectorAll<HTMLElement>(
-    ".mermaid-diagram[data-pending='true']"
-  );
-  for (const el of pending) {
-    const code = decodeURIComponent(el.getAttribute("data-code") ?? "");
-    if (!code) continue;
-    el.setAttribute("data-pending", "false");
-    const id = `mermaid-${++mermaidIdCounter}`;
-    try {
-      const { svg } = await mermaid.render(id, code);
-      el.innerHTML = svg;
-    } catch (err: unknown) {
-      const msg = err instanceof Error ? err.message : String(err);
-      el.innerHTML = `<pre class="mermaid-error">图表渲染失败:\n${msg}</pre>`;
-    }
-  }
+  await renderMermaidDiagrams(contentRef.value);
 }
 
 // 流式输出结束后（streaming → false）渲染图表
@@ -110,7 +61,7 @@ watch(
   async (isStreaming) => {
     if (!isStreaming) {
       await nextTick();
-      renderMermaidDiagrams();
+      renderPendingDiagrams();
     }
   }
 );
@@ -119,7 +70,7 @@ watch(
 onMounted(async () => {
   if (!props.message.streaming) {
     await nextTick();
-    renderMermaidDiagrams();
+    renderPendingDiagrams();
   }
   window.addEventListener("message", handleHtmlPreviewMessage);
 });
@@ -172,147 +123,6 @@ function handleMarkdownClick(event: MouseEvent) {
   }
 }
 
-// ============ HTML 预览块生成 ============
-
-/**
- * 为 HTML 代码块生成"源码 + iframe 预览"双视图结构。
- * iframe 使用 sandbox + srcdoc 隔离执行，toolbar 按钮的点击由
- * handleMarkdownClick 事件委托处理（不用内联 onclick，会被 CSP 拦截）。
- */
-function buildHtmlPreviewBlock(code: string): string {
-  const highlighted = hljs.highlight(code, { language: "html" }).value;
-
-  // 如果模型只输出了 HTML 片段（非完整文档），补全基础框架以确保正确渲染
-  const isFullDoc = /^\s*<!doctype/i.test(code) || /^\s*<html[\s>]/i.test(code);
-  const docHtml = isFullDoc
-    ? code
-    : `<!DOCTYPE html><html><head><meta charset="utf-8"><style>body{margin:16px;font-family:system-ui,sans-serif;line-height:1.6}</style></head><body>${code}</body></html>`;
-
-  // sandbox 不含 allow-same-origin（Issue #55：曾允许 iframe 内脚本靠
-  // allow-same-origin 逃出沙箱读父页面 localStorage）。iframe 因此是不透明源，
-  // 父页面无法用 contentDocument 读高度，改为在 iframe 内加载一个外部脚本文件，
-  // 通过 postMessage 主动上报高度（父页面处理见 handleHtmlPreviewMessage）。
-  // 必须是外部文件而不是内联 <script>：应用的 CSP 是精确哈希白名单（不含
-  // unsafe-inline/unsafe-hashes），且 srcdoc 文档会继承父页面 CSP，内联脚本
-  // 一律会被拦截静默失效；同源静态文件走 src= 加载则天然匹配 script-src 'self'。
-  // 闭合标签拆成两段字符串，避免在 .vue 的 <script> 块里出现完整闭合标签字面量。
-  const reporterTag = '<script src="/html-preview-reporter.js"></' + "script>";
-  const docWithReporter = /<\/body>/i.test(docHtml)
-    ? docHtml.replace(/<\/body>/i, `${reporterTag}</body>`)
-    : docHtml + reporterTag;
-
-  // srcdoc 属性值内只需转义 & 和 "
-  const srcdoc = docWithReporter.replace(/&/g, "&amp;").replace(/"/g, "&quot;");
-
-  // 按钮不再用内联 onclick（同样会被 CSP 拦截），改为在 handleMarkdownClick 里
-  // 用事件委托处理 .html-preview-toggle 点击。
-  return (
-    `<div class="html-preview-block">` +
-    `<div class="html-preview-toolbar">` +
-    `<button class="html-preview-toggle">预览效果</button>` +
-    `<span class="html-lang-badge">HTML</span>` +
-    `</div>` +
-    `<pre class="html-source-pre"><code class="hljs language-html">${highlighted}</code></pre>` +
-    `<iframe class="html-preview-frame" srcdoc="${srcdoc}" ` +
-    `sandbox="allow-scripts allow-modals allow-forms"></iframe>` +
-    `</div>`
-  );
-}
-
-// ============ Mermaid 预览块生成 ============
-
-/**
- * 为 Mermaid 代码块生成"图表预览 + 源码"双视图结构。
- * 默认展示渲染后的图表（show-diagram 类），点击按钮切换。
- * 图表占位元素 data-pending="true"，由 renderMermaidDiagrams() 在 DOM 就绪后填充。
- */
-function buildMermaidPreviewBlock(code: string): string {
-  const highlighted = hljs.highlight(code, { language: "plaintext" }).value;
-  // encodeURIComponent 保证特殊字符在 data-* 属性里安全传递
-  const encoded = encodeURIComponent(code);
-
-  // 按钮不再用内联 onclick（CSP 会拦截），改为在 handleMarkdownClick 里事件委托处理。
-  return (
-    `<div class="mermaid-preview-block show-diagram">` +
-    `<div class="mermaid-toolbar">` +
-    `<button class="mermaid-toggle">查看源码</button>` +
-    `<span class="mermaid-lang-badge">Mermaid</span>` +
-    `</div>` +
-    `<pre class="mermaid-source-pre"><code class="hljs">${highlighted}</code></pre>` +
-    `<div class="mermaid-diagram" data-pending="true" data-code="${encoded}"></div>` +
-    `</div>`
-  );
-}
-
-// ============ Markdown 配置 ============
-
-// marked 默认把消息正文里出现的裸 HTML（不在围栏代码块内的）原样透传到输出，
-// 这里转义掉，防止 <script>/<img onerror>/<style> 等直接变成真实 DOM 节点。
-// 围栏代码块走的是下面的 code() 渲染器，不受影响。
-function escapeRawHtml(text: string): string {
-  return text
-    .replace(/&/g, "&amp;")
-    .replace(/</g, "&lt;")
-    .replace(/>/g, "&gt;")
-    .replace(/"/g, "&quot;")
-    .replace(/'/g, "&#39;");
-}
-
-// 不少模型（如 DeepSeek）输出 LaTeX 公式时用的是 \[...\] / \(...\) 定界符而非
-// $$...$$ / $...$。marked-katex-extension 只认后者；更麻烦的是 CommonMark 规则
-// 会把 "\[" "\]" 这类反斜杠+标点解析成转义序列（渲染为去掉反斜杠的纯标点），
-// 定界符在 marked 的分词阶段之前就已经丢失了。所以要在喂给 marked 之前，
-// 先把 \[...\] / \(...\) 正规化成 $$...$$ / $...$。
-function normalizeLatexDelimiters(text: string): string {
-  return text
-    .replace(/\\\[([\s\S]*?)\\\]/g, (_match, expr: string) => `$$${expr}$$`)
-    .replace(/\\\(([\s\S]*?)\\\)/g, (_match, expr: string) => `$${expr}$`);
-}
-
-// 使用单例确保 marked.use() 只调用一次（marked 修改全局实例）
-const markedInitPromise = (() => {
-  let promise: Promise<void> | null = null;
-  return () => {
-    if (!promise) {
-      promise = new Promise((resolve) => {
-        // 黑白设计系统下不用 KaTeX 默认的红色报错色，退化为普通灰字
-        marked.use(markedKatex({ throwOnError: false, errorColor: "#444444" }));
-        marked.use({
-          renderer: {
-            // 消息正文中的裸 HTML（块级或内联）一律转义为纯文本
-            html(html: string, _block?: boolean): string {
-              return escapeRawHtml(html);
-            },
-            // 处理所有围栏代码块（三参数形式是此版本 marked 的 Renderer 签名）
-            code(text: string, lang: string | undefined, _escaped: boolean): string {
-              // 取 info string 第一个单词作为语言标识（如 "html filename.html" → "html"）
-              const normalizedLang = ((lang ?? "").trim().split(/\s+/)[0] ?? "").toLowerCase();
-
-              if (normalizedLang === "html" || normalizedLang === "htm") {
-                return buildHtmlPreviewBlock(text);
-              }
-
-              if (normalizedLang === "mermaid") {
-                return buildMermaidPreviewBlock(text);
-              }
-
-              // 其余语言走 highlight.js 常规高亮
-              const language = hljs.getLanguage(normalizedLang) ? normalizedLang : "plaintext";
-              const highlighted = hljs.highlight(text, { language }).value;
-              return `<pre><code class="hljs language-${language}">${highlighted}</code></pre>`;
-            },
-          },
-        });
-        resolve();
-      });
-    }
-    return promise;
-  };
-})();
-
-// 初始化 marked 配置
-markedInitPromise();
-
 // ============ 计算属性 ============
 
 // 是否为用户消息
@@ -321,23 +131,9 @@ const isUser = computed(() => props.message.role === "user");
 // 是否为 AI 助手消息
 const isAssistant = computed(() => props.message.role === "assistant");
 
-// 渲染后的 Markdown 内容
-// 二次防线：即便 marked 的转义有疏漏，DOMPurify 仍会按白名单剥离危险标签/属性。
-// iframe/sandbox/srcdoc 是 buildHtmlPreviewBlock 生成的沙箱预览所必需的，显式放行。
-const renderedContent = computed(() => {
-  if (!props.message.content) return "";
-  const normalized = normalizeLatexDelimiters(props.message.content);
-  const html = marked.parse(normalized, { async: false, breaks: true }) as string;
-  return DOMPurify.sanitize(html, {
-    ADD_TAGS: ["iframe"],
-    ADD_ATTR: ["sandbox", "srcdoc"],
-    // srcdoc's value is a full HTML document (it legitimately contains
-    // "</style>" etc.), which DOMPurify's SAFE_FOR_XML close-tag probe
-    // otherwise treats as an attribute-escape attempt and strips outright.
-    // The uponSanitizeAttribute hook above is what actually keeps srcdoc.
-    SAFE_FOR_XML: false,
-  });
-});
+// 渲染后的 Markdown 内容（解析、净化、HTML/Mermaid 预览块生成全部在
+// utils/markdown.ts 的 renderMarkdown 里完成）
+const renderedContent = computed(() => renderMarkdown(props.message.content));
 
 // ============ 方法函数 ============
 
@@ -398,8 +194,8 @@ const handleCopy = async () => {
         <div
           ref="contentRef"
           class="markdown-content"
-          v-html="renderedContent"
           @click="handleMarkdownClick"
+          v-html="renderedContent"
         />
         
         <!-- Streaming indicator -->
